updated tests, improved naming of methods
