+++ test/test_board.py
@@ -40,6 +40,39 @@ class TestBoard(unittest.TestCase):
         current_player = board.get_current_player(total_marks_on_board)
         self.assertEqual("X", current_player)
 
+    def test_X_is_placed_in_index_0_with_input_1(self):
+        board = Board()
+        current_board = ["1", "2", "3", "4", "5", "6", "7", "8", "9"]
+        user_input = 1
+        total_marks_on_board = 0
+        current_player = board.get_current_player(total_marks_on_board)
+        board.assign_board_index_to_current_player_mark(
+            user_input, current_board, total_marks_on_board
+        )
+        self.assertEqual("X", current_player)
+
+    def test_O_is_placed_in_index_1_with_input_2(self):
+        board = Board()
+        current_board = ["X", "2", "3", "4", "5", "6", "7", "8", "9"]
+        user_input = 2
+        total_marks_on_board = 1
+        current_player = board.get_current_player(total_marks_on_board)
+        board.assign_board_index_to_current_player_mark(
+            user_input, current_board, total_marks_on_board
+        )
+        self.assertEqual("O", current_player)
+
+    def test_X_is_placed_in_index_2_with_input_3(self):
+        board = Board()
+        current_board = ["X", "O", "3", "4", "5", "6", "7", "8", "9"]
+        user_input = 3
+        total_marks_on_board = 2
+        current_player = board.get_current_player(total_marks_on_board)
+        board.assign_board_index_to_current_player_mark(
+            user_input, current_board, total_marks_on_board
+        )
+        self.assertEqual("X", current_player)
+
     def test_take_turns_marking_board_when_no_marks_and_X_is_going_first(self):
         board = Board()
         current_board = ["1", "2", "3", "4", "5", "6", "7", "8", "9"]
@@ -49,7 +82,7 @@ class TestBoard(unittest.TestCase):
         self.assertEqual(
             new_board_with_marks,
             board.take_turns_marking_board(
-                current_board, user_input, total_marks_on_board
+                user_input, current_board, total_marks_on_board
             ),
         )
 
@@ -62,7 +95,7 @@ class TestBoard(unittest.TestCase):
         self.assertEqual(
             new_board_with_marks,
             board.take_turns_marking_board(
-                current_board, user_input, total_marks_on_board
+                user_input, current_board, total_marks_on_board
             ),
         )
 
@@ -75,6 +108,6 @@ class TestBoard(unittest.TestCase):
         self.assertEqual(
             new_board_with_marks,
             board.take_turns_marking_board(
-                current_board, user_input, total_marks_on_board
+                user_input, current_board, total_marks_on_board
             ),
         )
